recovery: Display recovery version
Set the title lines to the afterlife version and build date and render
them between the logo and menu.

The "AfterLife Recovery" and "Android Fastboot" lines are removed
because the logo already says "Recovery" or "Fastbootd".

Also: Assume the logo and back icon are non-null.
ScreenRecoveryUI::SelectMenu already makes this assumption, and it
avoids questions about how to handle "back" if either the logo or
the back icon is null.

Change-Id: Ic7a0c42b84b3e78b8f805bf2d22e7c63f07c0309
Signed-off-by: Aditya Eka Pranata <aepranata@gmail.com>