recovery: Display recovery version
Set the title lines to the lineage version and build date and render
them between the logo and menu.

The "Lineage Recovery" and "Android Fastboot" lines are removed
because the logo already says "Recovery" or "Fastbootd".

Also: Assume the logo and back icon are non-null.
ScreenRecoveryUI::SelectMenu already makes this assumption, and it
avoids questions about how to handle "back" if either the logo or
the back icon is null.

Change-Id: I4d6ad3818c09f53a8e5e3caca6205abbbd4b3095
Signed-off-by: Aditya Eka Pranata <aepranata@gmail.com>